// types.ts
export enum EducationalStage {
  PRE_SCHOOL = 'PRE_SCHOOL',
  PRIMARY = 'PRIMARY',
  MIDDLE = 'MIDDLE',
  HIGH = 'HIGH',
}

export enum Subject {
  // Primary
  Arabic = 'اللغة العربية',
  French = 'اللغة الفرنسية',
  Math = 'الرياضيات',
  IslamicEducation = 'التربية الإسلامية',
  SocialStudies = 'الاجتماعيات',

  // Middle & High
  English = 'اللغة الإنجليزية',
  Physics = 'الفيزياء والكيمياء',
  NaturalSciences = 'علوم الحياة والأرض',
  ComputerScience = 'المعلوميات',
  Philosophy = 'الفلسفة',
}


export interface Teacher {
  id: string;
  name: string;
  loginCode: string;
  subjects: Subject[];
  salary?: number;
  assignments: {
    [level: string]: string[];
  };
}